Started modifying the algorithm to be a class #62
